Perform bit decomposition as assignment if offset is known. (#2612)
If we detect a bit decomposition pattern like `X + Y * 256 + Z * 65536 =
T`, the witgen inference returns a bit decomposition effect. But if the
value of `T` is a known number, we can actually extract the concrete
values for `X`, `Y` and `Z` at compile-time and continue solving with
this additional information.