Enable JIT for permutations (#2593)
Enable support for permutation constraints. The main problem with
permutation constraints is that they need a way to disable constraints
and thus this PR adds a feature to guess values for variables that have
not been assigned yet but are also not relevant in any of the
identities, for example because they are multiplied by a value that is
known to be zero.

---------

Co-authored-by: Georg Wiese <georgwiese@gmail.com>