Poseidon2 half output (#2514)
Of all uses we have for Poseidon2 (Merkle Tree, Sponge absorption,
Sponge squeeze), none use the full 8 field elements of the output.
Sponge absorption uses only the final 4, and Merkle Tree and Sponge
squeeze uses only the initial 4.

This change introduces a switch for the user to select what parts of the
final state to output (first half, second half, all of it or none of
it).

---------

Co-authored-by: chriseth <chriseth.github@gmail.com>