`QuadraticSymbolicExpression`: Detect Bit constraint (#2653)
Adds a quick-and-dirty way to return bit constraints when running
`QuadraticSymbolicExpression::solve()`, by looking for a specific
expression (`X * (X - 1)`). As far as I can see, this is orthogonal to
#2619, which would return a `Branch` effect instead of a range
constraint.

With this PR, we remove the need for "global range constraints". This
will help progressing on the new solver (#2651), which tries to rely
only on `QuadraticSymbolicExpression`. Then, we can have a bit
decomposition example even before implementing machine calls.

---------

Co-authored-by: chriseth <chris@ethereum.org>