Add simple solver for arithmetic circuits  (#2651)
I tried to implement the simplest possible solver as a starting point
for a solver that we can eventually use to simplify auto-precompiles (by
(1) trying to find concrete variable assignments already and (2) figure
out if two nodes can be computed using the same symbolic expression, and
hence can be collapsed).

I created #2652 to track follow-up work.

---------

Co-authored-by: Thibaut Schaeffer <schaeffer.thibaut@gmail.com>
Co-authored-by: chriseth <chris@ethereum.org>