Implement solver-based pilopt (#2673)
Adds a new `optimizer` module, which is similar but not identical to the
optimizer introduced in #2643. It should be possible to completely
remove pilopt once we implement the equivalence of the
`replace_linear_witness_columns` step there as well.

This PR does not require changes in `powdr-openvm`, but I opened
https://github.com/powdr-labs/powdr-openvm/pull/94 to demonstrate that
the tests pass.