Use in-liner in APC optimizer (#2695)
This PR has the following changes:
- In the optimizer, I added a call `replace_constrained_witness_columns`
and some logging.
- I removed the rule that a substitution is only executed if the
substituted variable is used, for two reasons:
- The check was expensive (e.g. on Keccak the in-liner took longer than
the solver...)
- I think it's fair to just remove these identities then. It changes the
statement being proven, but the statement is trivial: "I know some
values that can be computed from other values in the table via a
low-degree expression"

As seen in https://github.com/powdr-labs/powdr-openvm/pull/108, this
leads to moderate improvement in the Keccak benchmark.