Add checked version of `handle_bus_interaction` (#2704)
Gives the bus interaction handler a way to return an error. That should
be helpful in #2698, because there, we're explicitly looking for
witnesses that violate a constraint.

The user of the trait does not have to do anything, because there is a
default implementation which simply checks if the old and new range
constraints overlap. So `powdr-openvm` does not need to be updated.
However, it enables a simplification
(https://github.com/powdr-labs/powdr-openvm/pull/114).