Exhaustive Search (#2698)
Adds another exhaustive search step to the solver: For each set of
variables that appear together in an expression, it checks whether the
number of possible assignments is small. If yes, it tries out all of
them and tests whether one of them is a unique solution. If yes, it
continues from there.

See the added tests for examples where this can be useful.

This PR is an alternative to #2683. It should work in more cases and
does not introduce a new effect. Also, the logic is quite encapsulated:
the solver does not have to keep track of a new constraint type.

This PR also enables us to remove some of the heuristics with which we
currently set the opcode flags, see #2699.

See https://github.com/powdr-labs/powdr-openvm/pull/115 for numbers on
the Keccak benchmark.

---------

Co-authored-by: chriseth <chris@ethereum.org>