fix apc register memory optimizer bug (#2717)
The bug that happened was that the last send of a register memory bus
interaction was removed due to having multiplicity zero. In the end the
bus was left unbalanced. Therefore, we need to run the solver optimizer
before the reg mem opt, so that these multiplicities are removed in
pairs before the apc optimizer runs and keeps only the first receive and
last send, because those multiplicities need to match. This is the main
solution implemented here in `build`. I also added some checks that the
reg mem interactions need to be even per address and in total.

A side effect of this fix is that we can no longer easily remove the
timestamp columns of removed mem int receives, the constraint that
decomposes it into bytes, the decomposing columns, and the range checks
on them. The pattern matching becomes too brittle and would be even
hackier if we wanted to keep supporting it that way. That's what causes
the regression in the test, and all the code that is removed in this PR,
since we won't do this anymore. The ultimate fix is to rely on the
optimizer to remove these columns, since:
- The reg mem int receive is removed
- Then the timestamp column becomes a free variable in its constraint
- The constraint can be removed
- The limb decomposition columns become unused in polynomial constraints
and only appear in range checks
- Therefore the range checks and columns can also be removed

At this point the result should be the same as the pattern matching
alternative removed by this PR.
I anticipate that we'll also need a fixpoint loop because it may be the
case that we need a few iterations to propagate everything possible when
this is implemented.