Remove disconnected columns (#2724)
This PR is an alternative to #2721, also fixing the regression of #2717.
It actually leads to benefits beyond that.

The idea is to only keep columns that are connected to *stateful* bus
interactions (i.e., memory and the execution bridge), because this is
the only way to affect the state of the VM. This is actually a special
case of the `remove_trivial_bus_interactions` step (trivial bus
interactions are connected to no columns, and therefore removed), so I
removed that (and simplified the `ConcreteBusInteractionHandler`).


[This](https://gist.github.com/georgwiese/56d768e86bbf6def28260ca92f0b5180)
is a diff of the generated PIL of the Keccak APC. I spot-checked a few
instances of removed columns and there were often columns that were
range-checked but otherwise un-constrained.