Executor: Generate AIR proof input (#2752)
Continuing #2733, this PR moves most of
`PowdrChip::generate_air_proof_input` into the `Executor`. As a result,
it is more encapsulated (all its fields are private now), and the
interface is such that we should be able to use it in the `PlonkChip` as
well: `Executor::generate_witness` returns a matrix of all variable
values, for each call to the APC. In the `PowdrChip`, this is the
witness directly; in the `PlonkChip`, we can compute the witness from
this matrix.

This PR *mostly* moves code. I would prefer to leave any bigger changes
to a different PR, so that it doesn't get too entangled.