Allow non-numeric asignments. (#2756)
This is a preparatory pull request to allow assignments of anything that
is not a concrete number.

This is needed once we reason that two columns are equivalent.

I do not plan to use it for anything more fancy than `X = Y`, or maybe
`X = k * Y`, but maybe not even that.