Plonk witness generation (#2767)
Builds on #2765 (already approved)

Implements a basic algorithm for PlonK witness generation.

Has a lot of TODOs and the code isn't actually run yet, but I'd suggest
to merge it already and iterate on that.

The general algorithm is:
- Call `Executor::generate_witness()` to get all the values of the
variables in the original constraint systems (= what would have been
trace cells in `PowdrChip`)
- For each call:
  - For each gate:
- Set the selector columns (which are currently witness columns, should
eventually be preprocessed columns)
- Solve for temporary variables on the fly, assuming that there is at
most unknown when processing gates in order
- Set the witness columns (`a`, `b`, `c`), with their values either
coming from the result of `Executor::generate_witness()` or a temporary
variable.