Associated openvm type (#2812)
As an alternative to #2804 use a new trait with an associated type to
have a single generic parameter instead of two.
After this PR, only the openvm cli is specific to BabyBear, everything
else is generic and generic type annotations are removed (since all
types can now be infered thanks to the associated type).

One current drawback is that the number crate must depend on `p3-sdk`
and `p3-backend`, although it only needs the field implementation. A
solution to that would be to point to the p3 crates openvm points to,
instead of using the reexports.

Update: I was wrong on only the CLI being generic. The StdIn type
exposed by openvm is specific to BabyBear, so everything in `lib` is
still BabyBear specific.

---------

Co-authored-by: Georg Wiese <georgwiese@gmail.com>