Add nightly run for apcs (#2851)
This adds a job to our nightly run that runs all tests in the
powdr-openvm crate, including recursion.
Future ideas:
- Split the common nightly into 2. I tried to do this but couldn't get
it to work initially. TBD
- Since we don't run Halo2 anymore, we can actually run recursion for a
lot of the tests. Next I would suggest to run the guest_recursion test
enabled here in a PR test actually that also uses the self hosted runner
- Similarly to above, add a larger Keccak test to PR tests