Fix OVM Witgen (#2836)
Pair here: https://github.com/powdr-labs/openvm/pull/24

OVM tests for creating multiple APCs (introduced in PGO branch) fails.

This patches the error that we cannot skip witgen for the last
read/write access to memory.

Previously we thought this was the last access, but some APC has
timestamp increment as the last access, which is not a "real" access.

---------

Co-authored-by: schaeff <thibaut@powdrlabs.com>
Co-authored-by: Leo Alt <leo@powdrlabs.com>