Journal (#2852)
Introduces another abstraction layer above IndexedConstraintSystem that
tracks changes made to the constraint system.

The actual change tracking is not yet implemented, because the optimizer
has not been switched to ConstraintSystem yet, but it is already very
beneficial that there is no full access to the items in the constraint
system.