Refactor bus map retrieval (#2885)
We currently include the bus map in the powdr config. However, it is not
a parameter we can chose, but rather depends on the original vm. Move it
there, still returning the default one. Next step would be ideally
generating it from the vm config.

Also use `OriginalProgram` when possible, to keep the program and config
together.