Register 0 returns 0 (#2899)
We can actually encode the `x0` semantics in the memory bus interaction
handler.

Note that this still keeps a register access to `x0` that could be
removed, but now at least we're not allocating 4 witness columns for the
value *and* we might be able to derive more values at compile time.