Legacy Expression: Remove next reference (#2832)
Builds on #2890

This PR removes the "next" flag from `AlgebraicReference`. This requires
a few changes, because OpenVM's AIRs do use them. It should still be
removed, because we can't deal with it when generating APCs.

The solution is to introduce a new type that wraps `AlgebraicReference`:

https://github.com/powdr-labs/powdr/blob/1cee98dbba30051ce915566175499f8c28ee831b/openvm/src/utils.rs#L25-L30

Note that this also takes out the `is_{first,last,transition}`
reference, which we also can't handle when compiling APCs!

When exporting OpenVM's chips to PIL, we use the `OpenVmReference` type;
for everything else, we convert to `AlgebraicReference`. For this, we
need to make sure that we don't do that for blacklisted opcodes (because
they might contain next references).

---------

Co-authored-by: Leo <leo@powdrlabs.com>