Introduce Bus Interaction variables (#2900)
This PR enables us to remove an ugly heuristic: #2867

For context, this was the problem:
- In OpenVMs load/store chip, the memory interactions going to register
address space still include unknown columns in the address expression.
This is correct (we can't solve for these columns at compile time), but
we can narrow down the possible values for these columns. If you do
that, it turns our that for all possible assignments, the address
evaluates to the same constant value.
- #2866 added this type of reasoning to variables. But before this PR,
only witness columns are considered variables. The (register) memory
address is not stored in a witness column though.

This PR solves this by introducing variables for bus interaction fields.

When the solver is started, variables are wrapped with this type:

https://github.com/powdr-labs/powdr/blob/b23cfc5f70474c9ddacf35c15ffe2437419c6a3e/constraint-solver/src/solver/bus_interaction_variable_wrapper.rs#L12-L17
and bus interaction fields are replaced with `Variable::
BusInteractionField` instances. Additional constraints are added to bind
the new variable to the expression that was replaced.

In the end, the variables are unwrapped again and any bus interaction
field assignments are returned to the caller.

In the end, the solver is slower because of [this
addition](https://github.com/powdr-labs/powdr/blob/b23cfc5f70474c9ddacf35c15ffe2437419c6a3e/constraint-solver/src/solver/exhaustive_search.rs#L132-L148).
Measuring `time cargo test -r tests::keccak_machine_pgo -- --nocapture`
on my system, I get:
- 45s on `main`
- 98s on `introduce-bus-interaction-vars2` (this PR)
- 99s on `remove-is-load-heuristic` (#2867)
- 48s on `tweak-exhaustive-search-params` (#2908) after tweaking some
parameters