Docs for ovm assembly builder (#2893)
Depends on https://github.com/powdr-labs/powdr/pull/2876

Created this PR in response to comments:
https://github.com/powdr-labs/powdr/pull/2876#discussion_r2142388765

Currently, I composed the doc for each instruction such that it contains
the OVM doc for the adapter and core chips of each instruction. This has
the advantage of fully describing the constraints expected from each
instruction.

Because the doc should be repeated for each group of instructions, e.g.
(ADD and SUB), for an easy review, I'm not copying the docs to opcodes
that it should be repeated for yet, but eventually I would attempt to
use some macro schenanigans to avoid duplication as much as possible,
but in the worst case, repeat them for some opcodes.

P.S. They look pretty nice when you hover over the constructor function
with an IDE. :)