Extract bus map from sdk config (#2892)
Verbose, but working.
After this, the default config could be moved back into the
autoprecompiles repo to be used only in unit tests.

Update: not verbose anymore. We use `find_chip` with the periphery chips
that we are expecting. The exhaustive approach seemed like too much code
for the benefit: we would know if a new chip gets added, but we wouldn't
if a new bus gets added to an existing chip. So the current version is
good enough.