Use allow list for opcode (#2915)
Currently we use a black list for opcode that shouldn't appear in a BB
for APC. This method has the disadvantage of encountering many bugs due
to new opcodes appearing that we can't create APC for (whether due to
unsupported expression type or compile time memory overload).

With the new reth-benchmark, we identified a few custom opcodes (likely
manual precompiles) that get injected in an APC, which causes an
abnormal amount of memory usage up to the 260 gigs limit of our server.
This calls for a systematic way to use an allow list for opcodes so that
we are robust against future manual opcodes.

A few key changes:
1. Use a macro to derive the vector of allowed opcodes from public
constants.
2. Allow list = basic opcodes - hint opcodes + bigint branch opcodes

Co-authored-by: schaeff <thibaut@powdrlabs.com>