Cache even fewer APC for Pgo::Cell (#2920)
Because we know the number of costliest APCs to compute at compile time,
we only need to cache the top `APC + APC_SKIP` costliest APCs for
Pgo::Cell mode instead of all of them. This is achieved by:
1. Maintaining a binary heap data structure which tracks the cost of
blocks and if we are over the max number of cache, insert the new cache
and eject the lowest costed block if the former is not the minimum
costed block.
2. Still keeping APC generation parallel by adding mutex guard around
the binary heap which is used for cost ordering as well as the map of
APC caches.