[PR for Review] Detect and count non-Powdr columns for cell PGO max column use (#2955)
Pgo input is now "max total columns" instead of "max apc columns",
because there can be many non APC columns. To be exact, roughly ~580 for
RV32 chip set and ~15K in the set used in `reth-benchmark`. Tested and
works for Keccak.

This PR calculates the total non APC columns number, logs them by chip
in case we interested in what airs they belong to, and then subtract the
non APC total from the max total the user supplies to PGO to obtain the
max APC columns, which is then used in the KnapSack algorithm to limit
APC generation in Pgo::Cell mode.

This PR's method is a bit of simplification, we are not sure if all
chips in the original config's chip inventory get eventually proven.
There are three scenarios (in order of ascending restrictiveness):
1. All columns in original inventory stay in. 
2. Those that get executed stay in. 
3. Those that get proven stay in. 

Scenario 3 is more restrictive than scenario 2, because we need all
dummy chips executed for witgen but not necessarily all dummy columns
for proving.