Update patch (#2961)
Simple one. Update patch so that it matches up exactly to the git
dependencies.

Before this PR, I had many compilation errors (cuz we are missing a
couple local crates) when syncing up patches during local development,
so might as well push to main so others don't need to fix up in the
future ;)