Refactor `QuadraticSymbolicExpression` (#2923)
This PR refactors `QuadraticSymbolicExpression`: I removed the reference
to
[`SymbolicExpression`](https://github.com/powdr-labs/powdr/blob/f60e9d9f699810b565b077911c3836bcb1f9a356/constraint-solver/src/quadratic_symbolic_expression.rs#L70).
It can represent values that are known at runtime, represented as an
expression. We don't need this in the context of APC: Variables are
either unknown or known at compile time, and therefore can be
represented as a `FieldElement` instead.

The idea is to introduce [this
trait](https://github.com/powdr-labs/powdr/blob/4989be08f343ea4b8e8d8a8e30a97b247b9b1a71/constraint-solver/src/runtime_constant.rs#L11),
which is implemented by `SymbolicExpression` and any `T: FieldElement`.
This way, we can continue to use the solver with
`QuadraticSymbolicExpression<SymbolicExpression<T, V>, V>` (equivalent
to the current `QuadraticSymbolicExpression<T, V>`) in the old JIT
pipeline and use the simpler `QuadraticSymbolicExpression<T, V>` in the
APC pipeline.

---------

Co-authored-by: chriseth <chriseth.github@gmail.com>