Avoid collecting into pc vector (#2989)
When we generate the PGO data, we first collect all visited pcs in
execution order which is linear in the execution length.
Then we reduce this to a map of unique PC to number of occurrences,
which is linear in the program size.

This PR removes the first collection, generating the map directly.

Update: since for reth we do ~300 million additions to the hashmap, this
PR also switches to using a vector to keep track of the count as opposed
to a hashmap. Not to be confused: before this PR the vector is the
sequence of all pcs visited. After this PR, it is the number of times a
pc was visited, by pc.

Other update:
copilot made a good point about switching away from
`Arc<Mutex<Vec<u32>>>` to `Arc<Vec<AtomicU32>>`. Done.