Improve OVM compilation tests (#3026)
In general, it makes more sense to always test:
- The full `AirMetrics` (including log up and preprocessed columns),
instead of just the main columns.
- The sum of `AirMetrics` for all APCs used, instead of just the first
APC in a testing profile (because of wasted data, and we can always just
create one APC if we only want to test one).

These won't add any cost to the testing CI run, because all data are
already there but we just aren't testing them fully yet.