Store column count. (#3066)
In order to get a feel for the optimizer, I think it is good to also
output the column count. It is rather difficult to get the column count
before optimization, but in the end we are only interested in the effect
of a new optimization step anyway, so I think this is fine.

---------

Co-authored-by: Georg Wiese <georgwiese@gmail.com>