Extract reachability. (#3069)
This change extracts the reachability algorithm and adds a new feature
that we will need for the seqz optimization: Find all variables that are
connected to a stateful bus interaction in any way that does not go
through a certain set of variables. The idea will be that we know of a
more efficient way to compute a certain relation between several input
variables and one output variable. This reachability routine can be used
to determine which constraints are used to compute the relation and can
be replaced by the more efficient routine.