Only transform non-affine bus interactions. (#3101)
Bus interaction variables were introduced in order to be able to derive
knowledge about complex expressions used in bus interactions. The system
is quite good in handling affine expressions, though, so in order to
keep the number of variables (and constraints) down, we only replace
non-affine expressions in bus interactions.