Remove (some) redundant range constraints (#3169)
While checking whether the solver already knows about a range
constraint, it also tries to add those that are new. This currently only
works if the expression is just a reference to a single variable.

It does not lead to any improvements in OpenVM, but I think it still
makes sense.