Update Bus Interaction handler for the variable range checker bus (#3178)
Extracted from #3130 

See [this
comment](https://github.com/powdr-labs/powdr/pull/3130/files#r2268175800):
> If the bit value is not in range, `handle_variable_range_checker` will
return the expected range, and `handle_bus_interaction_checked` will
intersect it with the previous range constraint (a known number), notice
the ranges don't overlap, and return an error.

As part of #3130, it might happen that the brute-force algorithm tries
to set the `bits` field to an out-of-range value. As mentioned in the
comment above, this would anyway result in an error if
`handle_bus_interaction_checked` is used, so panicking here is overly
strict.