Use a different cache for test_apc (#3179)
The `illegal instruction` error is back on nightly, see here:
https://github.com/powdr-labs/powdr/actions/runs/16954946528/job/48054989302

We suspect it could because `test_apc` uses the same cache as the normal
tests, but that cache has its build done in a different machine so the
dependencies could have instructions that cannot run on the server. If
this is the case, it happens most likely from the C/C++ dependency.

Anyway, we don't know that for sure, but this PR seems to fix it, see
here
https://github.com/powdr-labs/powdr/actions/runs/16944315837/job/48021334148.