Remove duplicate factors. (#3205)
This new optimizer step removes duplicate factors in a single algebraic
constraint.

It was detected by the assertion that is now commented, which assumed
that if all factors of one constraint also appear in another constraint
then the number of factors is also less or equal the number of factors
in the other constraint.

By running the duplicate factor removal right before that other step
this assumption gets true.