Nightly hot fix: Move tests after benchmarks (#3222)
Nightly has been failing for a while. We should fix that, but by moving
ignored the tests after the benchmarks, we should at least get nightly
benchmark results again (and test our most important guest programs).

I think this is also a good idea in general, because we often manually
run nightly to get the latest benchmarks on a particular branch. This
way, we get results earlier.

I started a manual run on this branch:
https://github.com/powdr-labs/powdr/actions/runs/17212171321